LayoutTests/media/media-source/media-source-h264-bframes-lowfps.html

<!DOCTYPE html>
<html>

<head>
  <title>media-source-h264-bframes-lowfps</title>
  <script src="media-source-loader.js"></script>
  <script src="../video-test.js"></script>
  <script>
    var loader;
    var source;
    var sourceBuffer;

    function loaderPromise(loader) {
      return new Promise((resolve, reject) => {
        loader.onload = resolve;
        loader.onerror = reject;
      });
    }

    window.addEventListener('load', async event => {
      try {
        findMediaElement();
        loader = new MediaSourceLoader('content/test-h264-bframes-lowfps-manifest.json');
        await loaderPromise(loader);

        source = new MediaSource();
        run('video.src = URL.createObjectURL(source)');
        await waitFor(source, 'sourceopen');
        waitForEventAndFail('error');

        run('sourceBuffer = source.addSourceBuffer(loader.type())');
        run('sourceBuffer.appendBuffer(loader.initSegment())');
        await waitFor(sourceBuffer, 'update');

        consoleWrite('Append media segments.');
        for (let i = 0; i < loader.mediaSegmentsLength(); i++) {
          run(`sourceBuffer.appendBuffer(loader.mediaSegment(${i}))`);
          await waitFor(sourceBuffer, 'update');
        }

        run('source.endOfStream()');
        run('video.play()');
        await waitFor(video, 'play');
        await waitFor(video, 'ended');

        endTest();
      } catch (e) {
        failTest(`Caught exception: "${e}"`);
      }
    });
  </script>
</head>

<body>
  <div>
    media-source-h264-bframes-lowfps
  </div>
  <video controls></video>
</body>

</html>
