LayoutTests/media/media-source/media-source-h264-bframes-lowfps-expected.txt

media-source-h264-bframes-lowfps

RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer(loader.type()))
RUN(sourceBuffer.appendBuffer(loader.initSegment()))
EVENT(update)
Append media segments.
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(0)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(1)))
EVENT(update)
RUN(source.endOfStream())
RUN(video.play())
EVENT(play)
EVENT(ended)
END OF TEST
